permutation header
